% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geno2r-package.R
\docType{package}
\name{geno2r-package}
\alias{geno2r}
\alias{geno2r-package}
\title{geno2r: Read Genotype Data To R}
\description{
In this package, you will find functions that allow you to read genotype data into the R environment.
}
\author{
\strong{Maintainer}: Jen-Hsiang Ou \email{oumark.me@outlook.com}

}
\keyword{internal}
